backend/requirements.txt
fastapi>=0.109.0
uvicorn==0.24.0
pydantic>=2.11.0
python-dotenv==1.0.0
openai>=1.10.0
faiss-cpu
numpy==1.26.2
python-multipart==0.0.6
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
langchain
langchain-community
langchain-core
langgraph
langchain-openai
sentence-transformers